Add an implementation of snprintf and vsnprintf for windows.
MSVC doesn't have snprintf so we have to implement it ourselves using
_vsnprintf_s and _vscprintf. This is in ExtraWrappers.h. We also provide
a definition of vsnprintf (which exists in MSVC but with broken
semantics).

This commit also renames ExtraSyscalls.h to ExtraWrappers.h on linux so
that the names match.

All calls are named ds2_{,v}snprintf, so that the code is the same
everywhere without #ifdefs.