Fix an off-by-one error when parsing ldmia.w
ARM instruction "ldmia.w" has a reglist, which is always non-empty.
If the reglist contains register PC, we should use the "size of
the reglist minus 1" as the offset of the data for register PC.
The previous code uses the "size of the reglist" as the offset,
which is off by one.