Avoid race conditions with _resumeSession in DebugSessionImpl.
When we stop a process and return from `onResume`, we set _resumeSession
to nullptr. If the process outputs some data to one of its standard
streams but the handler thread does not get scheduled until we return
from onResume, when the thread gets to run, _resumeSession will be null.

This commit avoids this scenario by locking accesses to this variable.