Handle Thread Exit Cleanly
Previously DS2 updates a thread and then checks for _stopInfo.event.
This leads to a problem where a thread placeholder remains for too long
and when updateState() invokes ReadStat, there is no such thread to be
found in /proc/.... This change simply checks _stopInfo.event first and
return early if the thread has already exited.