Add ENDIAN defines in Base.h. Fixes issue 49.
It turns out that __BYTE_ORDER__ and friends are not defined
by MSVC.  This lead to a bug, because expressions like
  #if __BYTE_ORDER__ == __ORDER_BIG_ENDIAN__
would evaluate as true, regardless of actual endianness.

With this change we stop using the aforementioned endianness
defines on Windows, and instead simply assume little endian.

This seems to be a safe assumption for modern Windows platforms.