Add wrapper for builtin bits functions
Targets like Windows ARM don't support this kind of builtins, thus
we fallback to manual calculations.
Whenever we can use, for example, __builtin_popcount, we use it for
chunks of #bits(int) bits just to make it more generic and avoid having a big
switch to call the corresponding builtins for each int size.