Allow several options to be set by env vars.
These options are
DS2_HOST
DS2_PORT
DS2_RUN_MODE
DS2_PROGRAM

The addOption method from OptParse can now receive an optional envVarName.
The value from the env var is only used it the option hasn't been set
by the normal arguments, i.e. env vars have lower precedence.

The only bad thing is that DS2_RUN_MODE can only be used if no arguments
are passed. This is fixable... but it would be a major refactor that
might not be necessary

I tested it on both windows and linux