Do not touch `lr` when removing the thumb bit from `pc` on winphone.
Windows on ARM does this weird thing of setting the lowest bit of `pc`
in the `CONTEXT` structure (output structure of `GetThreadContext()`)
even though that register doesn't actually have the lowest bit set in
the CPU; only the thumb bit in `CPSR` indicates we are running in thumb
mode.

`lr` on the other hand does have the lowest bit set if the caller was a
thumb function so we shouldn't remove it.