Extend HwBPMs _sites map to include  address values
For the software implementation of readonly hardware breakpoints, we
need to keep an up-to-date map of the values at memory addresses. That
way when our read|write breakpoint is triggered, we can check to see if
the value at that memory address has changed before deciding whether or
not to halt at that watchpoint

The downside with this approach is if one re-writes the existing value
at one of these hardware readonly watchpoints, this implementation will
halt there, thinking the value was read from (since the address was
halted by our read|write breakpoint and the value is unchanged), instead
of correctly ignoring that halt as a write.