Implement fast memory read/write on Linux
Instead of sticking with ptrace for every POSIX system, we can use
process_vm_readv and process_vm_writev on Linux. This will allow us to
do our memory operation with one syscall instead of many 4-byte reads
with ptrace.

If these fail, we still have a fallback to use the ptrace mechanism.