Stop using __has_attribute and __has_builtin
All the attributes and builtins we use are supported by the clang/gcc
versions we care about so there's no need for extra fancy checks for
those.

The mandatory ones have implicit errors that will trigger if somebody
builds the source with an unsupported compiler. The optional ones
(e.g.: DS2_ATTRIBUTE_FORMAT) will simply be defined to nothing and will
be transparently dropped.