Add a hacky patch to make android-arm tests run properly.
There are a large number of tests that rely on getPlatform() in this
python file. However, getPlatform() raises an exception because
the call to GetTriple() returns none (for some reason). We don't test
on any platforms that aren't linux or based on linux (e.g.  Android), so this
shouldn't actually break anything currently. It should be an active goal
of this project to remove this patch quickly.

The commit deleting this patch should contain a postmortem explainging
why this patch was needed, and how it can be fixed. If the cause is not
ds2's fault but lldb's, we should remove this and carry a better patch
on top of lldb. Assuming that's the case, we should also
upstream said patch if it's not fixed already.