Implement Linux::PTrace::getEventMessage
Upon a tracee calling clone(3) the tracer can call ptrace(3) with the
PTRACE_GETEVENTMSG flag to receive the tid of the newly cloned/spawned
thread.

Implement a method, Linux::PTrace::getEventMesssage that can be called
after a clone event that sets the value of clonedTID to the newly
cloned child's tid.