Modify Linux::Process::wait() to account for a race condition
It is possible for the inferior to experience a (roughly) simultaneous
SIGTRAP from stepping while also receiving a SIGTRAP/SIGSTOP pair from a
clone(3) call. Depending on the receiving order of the signals, ds2
could handle the situation improperly and miss out on the existance of
the cloned thread.

Modify the processing logic to check if we are simultaneously
experiencing a thread being stepped while that thread also calls
clone and, if so, handle the cloned child at within the function instead
of depending on the normal waitpid loop.