Fix two bugs involving the timing of waitpid events
Given a scenario where a inferior thread is being continuously single
stepped and hits a clone(2) call, the kernel will have three messages to
send the tracer:

1. The child thread has been created and is stopped.
2. The parent thread was stopped due to the single step.
3. The parent thread was stopped due to the clone event.

Previously, order of these events was always started with #2 or #3. But
on Android-ARM64, we are now seeing event #1 sent first.

So the first fix is to implement a check to see if the spawned thread
has already been waited on. In this case, we just call `step` and
continue waiting.

Antoher error existed where there was a significant delay being the
kernel sending event #3 and event #1 to the point where calling
waitpid(2) would miss #1. This would be ackknowledged as a failure.
Shortly after the event would be sent at which point we would no longer
be looking for it.

The second fix would be to wait and hang until the singal is sent since
the kernel has guaranteed that a clone event has already happened and
the fix mentioned above guarantees that we have yet to seen it, thus we
should be waiting for it.