Set the lldb packet timeout high while testing
Running the test suite on an emulator over ADB through python in a
container in a vm multithreaded on a ci server causes timeout issues.
lldb detects a timeout and tries to recover by sending `qC` to the
platform ds2 and expects a `QC:<some-tid>` in response. Sending `qC`
(which queries for the current thread info) doesn't make sense here in
since the platform ds2 never actually debugs a process. Thus the
platform ds2 correctly returnss `""`. Since we experience timeouts
where lldb's timeout recovery system fails:

Introduce a `DS2TESTING` macro to be defined while running on CircleCI
and drastically increase the timeout timer when defined.